added some features in drawer, added apis to handle change password, verify otp and update avatar

+++ lib/widgets/custom_drawer.dart
@@ -1,15 +1,18 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:get/get.dart';
 import 'package:mathquiz_mobile/config/color_const.dart';
 import 'package:mathquiz_mobile/config/demension_const.dart';
 import 'package:mathquiz_mobile/config/routes.dart';
 import 'package:mathquiz_mobile/features/auth/data/local_data_controller.dart';
+import 'package:mathquiz_mobile/features/auth/getx/auth_controller.dart';
 import 'package:mathquiz_mobile/features/drawer/drawer_controller.dart';
 
 import '../config/media_query_config.dart';
 
 class CustomDrawer extends StatelessWidget {
   final CustomDrawerController? controller;
+
   const CustomDrawer({
     this.controller,
     super.key,
@@ -20,45 +23,97 @@ class CustomDrawer extends StatelessWidget {
     SizeConfig().init(context);
     final LocalDataController localDataController =
         Get.put(LocalDataController());
-    // final screenHeight = MediaQuery.of(context).size.height;
+    final AuthController authController = Get.put(AuthController());
     return Drawer(
       width: 0.85 * SizeConfig.screenWidth!,
       child: Obx(() => localDataController.isLoading.isFalse
           ? Column(
               mainAxisAlignment: MainAxisAlignment.start,
               children: [
-                Container(
-                  width: double.infinity,
-                  height: 170,
-                  color: ColorPalette.primaryColor,
-                  child: Padding(
-                    padding: const EdgeInsets.only(
-                        left: kDefaultPadding, top: kDefaultPadding),
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        const CircleAvatar(
-                          radius: 30,
-                          backgroundColor: ColorPalette.backgroundColor,
-                          child: Icon(
-                            Icons.person_2_outlined,
-                            color: ColorPalette.primaryColor,
-                          ),
+                Column(
+                  children: [
+                    Container(
+                      width: double.infinity,
+                      height: 170,
+                      color: ColorPalette.primaryColor,
+                      child: Padding(
+                        padding: const EdgeInsets.only(
+                            left: kDefaultPadding, top: kDefaultPadding),
+                        child: Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: [
+                            GestureDetector(
+                              onTap: () async =>
+                                  await authController.pickImage(),
+                              child: Stack(
+                                children: [
+                                  Padding(
+                                      padding: const EdgeInsets.all(2),
+                                      child: CircleAvatar(
+                                        radius: 30,
+                                        backgroundColor:
+                                            ColorPalette.backgroundColor,
+                                        child: !authController
+                                                .isUploadingAvatar.value
+                                            ? localDataController.clientImageUrl
+                                                    .value.isEmpty
+                                                ? const Icon(
+                                                    Icons.person_2_outlined,
+                                                    color: ColorPalette
+                                                        .primaryColor,
+                                                  )
+                                                : ClipOval(
+                                                    child: Image.network(
+                                                      localDataController
+                                                          .clientImageUrl.value,
+                                                      fit: BoxFit.cover,
+                                                      width: 60.0,
+                                                      height: 60.0,
+                                                    ),
+                                                  )
+                                            : const Center(
+                                                child: SizedBox(
+                                                    height: 25,
+                                                    width: 25,
+                                                    child:
+                                                        CircularProgressIndicator()),
+                                              ),
+                                      )),
+                                  Positioned(
+                                      right: 0,
+                                      top: 0,
+                                      child: Container(
+                                        decoration: const BoxDecoration(
+                                          shape: BoxShape.circle,
+                                          color: ColorPalette.backgroundColor,
+                                        ),
+                                        child: const Padding(
+                                          padding: EdgeInsets.all(2.0),
+                                          child: Icon(
+                                            Icons.edit_outlined,
+                                            size: 15,
+                                          ),
+                                        ),
+                                      ))
+                                ],
+                              ),
+                            ),
+                            const SizedBox(
+                              height: 10,
+                            ),
+                            Text(
+                              localDataController.clientFullName.value
+                                  .toUpperCase(),
+                              style: const TextStyle(
+                                  color: Colors.white,
+                                  fontSize: 20,
+                                  fontWeight: FontWeight.w600),
+                            ),
+                          ],
                         ),
-                        const SizedBox(
-                          height: 10,
-                        ),
-                        Text(
-                          localDataController.clientFullName.value
-                              .toUpperCase(),
-                          style: const TextStyle(
-                              color: Colors.white,
-                              fontSize: 20,
-                              fontWeight: FontWeight.w600),
-                        ),
-                      ],
+                      ),
                     ),
-                  ),
+                  ],
                 ),
                 Expanded(
                   child: Padding(
@@ -74,7 +129,32 @@ class CustomDrawer extends StatelessWidget {
                             controller!.closeDrawer();
                             Get.toNamed(Routes.examHistoryScreen);
                           },
+                          title: const Text('Thông tin cá nhân'),
+                        ),
+                        ListTile(
+                          contentPadding: const EdgeInsets.all(0),
+                          onTap: () {
+                            controller!.closeDrawer();
+                            _showChangePasswordDialog(context);
+                          },
+                          title: const Text('Đổi mật khẩu'),
+                        ),
+                        ListTile(
+                          contentPadding: const EdgeInsets.all(0),
+                          onTap: () {
+                            controller!.closeDrawer();
+                            Get.toNamed(Routes.examHistoryScreen);
+                          },
                           title: const Text('Lịch sử làm bài'),
+                        ),
+                        const Divider(),
+                        ListTile(
+                          contentPadding: const EdgeInsets.all(0),
+                          onTap: () {
+                            controller!.closeDrawer();
+                            Get.toNamed(Routes.examHistoryScreen);
+                          },
+                          title: const Text('Giới thiệu'),
                         ),
                         const Spacer(), // Add Spacer to push the button to the bottom
                         Row(
@@ -85,12 +165,46 @@ class CustomDrawer extends StatelessWidget {
                               width: 150,
                               child: ElevatedButton(
                                   onPressed: () async {
-                                    await localDataController
-                                        .deleteClientEmail();
-                                    await localDataController
-                                        .deleteClientFullName();
-                                    await localDataController.deleteClientId();
-                                    Get.offAllNamed(Routes.loginScreen);
+                                    // Show a confirmation dialog
+                                    bool? confirm = await showDialog<bool>(
+                                      context: context,
+                                      builder: (BuildContext context) {
+                                        return AlertDialog(
+                                          title: const Text('Đăng xuất'),
+                                          content: const Text(
+                                              'Bạn có chắc chắn muốn đăng xuất không?'),
+                                          actions: [
+                                            TextButton(
+                                              onPressed: () {
+                                                Navigator.of(context)
+                                                    .pop(false);
+                                              },
+                                              child: const Text('Hủy'),
+                                            ),
+                                            TextButton(
+                                              onPressed: () {
+                                                Navigator.of(context).pop(true);
+                                              },
+                                              child: const Text(
+                                                'Đăng xuất',
+                                                style: TextStyle(
+                                                    color: Colors.red),
+                                              ),
+                                            ),
+                                          ],
+                                        );
+                                      },
+                                    );
+
+                                    if (confirm == true) {
+                                      await localDataController
+                                          .deleteClientEmail();
+                                      await localDataController
+                                          .deleteClientFullName();
+                                      await localDataController
+                                          .deleteClientId();
+                                      Get.offAllNamed(Routes.loginScreen);
+                                    }
                                   },
                                   child: const Text('Đăng xuất')),
                             ),
@@ -112,6 +226,96 @@ class CustomDrawer extends StatelessWidget {
           : const Center(
               child: CircularProgressIndicator(),
             )),
+    );
+  }
+
+  void _showChangePasswordDialog(BuildContext context) {
+    final TextEditingController oldPasswordController = TextEditingController();
+    final TextEditingController newPasswordController = TextEditingController();
+    final TextEditingController confirmPasswordController =
+        TextEditingController();
+    final authController = Get.put(AuthController());
+    showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: const Text('Đổi mật khẩu'),
+          content: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              TextField(
+                controller: oldPasswordController,
+                decoration: const InputDecoration(
+                  labelText: 'Mật khẩu cũ',
+                ),
+                obscureText: true,
+              ),
+              TextField(
+                controller: newPasswordController,
+                decoration: const InputDecoration(
+                  labelText: 'Mật khẩu mới',
+                ),
+                obscureText: true,
+              ),
+              TextField(
+                controller: confirmPasswordController,
+                decoration: const InputDecoration(
+                  labelText: 'Xác nhận mật khẩu mới',
+                ),
+                obscureText: true,
+              ),
+            ],
+          ),
+          actions: [
+            TextButton(
+              onPressed: () {
+                Navigator.of(context).pop();
+              },
+              child: const Text('Hủy'),
+            ),
+            SizedBox(
+              height: 35,
+              width: 150,
+              child: ElevatedButton(
+                onPressed: () async {
+                  String oldPassword = oldPasswordController.text;
+                  String newPassword = newPasswordController.text;
+                  String confirmPassword = confirmPasswordController.text;
+
+                  if (newPassword.isEmpty ||
+                      oldPassword.isEmpty ||
+                      confirmPassword.isEmpty) {
+                    Get.snackbar(
+                        'Lỗi', 'Vui lòng nhập đầy đủ các trường yêu cầu');
+                  } else if (newPassword == confirmPassword) {
+                    await authController.changePassword(
+                        oldPassword, newPassword);
+                    Navigator.of(context).pop();
+                  } else {
+                    Get.snackbar('Lỗi', 'Mật khẩu không khớp');
+                  }
+                },
+                child: Obx(
+                  () => authController.isLoading.value
+                      ? const Center(
+                          child: SizedBox(
+                            height: 20,
+                            width: 20,
+                            child: CircularProgressIndicator(
+                              color: Colors.white,
+                            ),
+                          ),
+                        )
+                      : const Text(
+                          'Đổi mật khẩu',
+                          style: TextStyle(color: Colors.white),
+                        ),
+                ),
+              ),
+            ),
+          ],
+        );
+      },
     );
   }
 }
